Cleanup pass: stale pointers, prose-code drift, and small redundancies
The replication handoff no longer claims its draft is unmerged (it is
section 24). The living docs that counted 6''' at three legs now count four
and cite the measurement: what-gd-builds item 1 carries the section 25-26
narrowing, item 2 records that the inequality-native thermometer family was
run and died (the empirical arm of its theory-grounds closure), and section
23's closing line points forward to section 26. probe_designedgates's
docstrings caught up with the code (the gates are AND/NAND, not AND/OR, and
build_hadamard now explains why only mixn survives the empty-fact check);
section 26's variant count matches what was actually tried. Sparsity in
section 25 names the embedding weights specifically; the reply loses a
doubled ten-kilobits; two redundant int(round()) casts go.

+++ probe_ratedistortion.py
@@ -362,7 +362,7 @@ class Probe:
             tag = f"tie_q{q}"
             if self.skip("tieflip", tag):
                 continue
-            n_flip = int(round(q * self.pattern_t.size))
+            n_flip = round(q * self.pattern_t.size)
             mask = np.zeros(self.pattern_t.size, dtype=bool)
             mask[order[:n_flip]] = True
             pat = self.pattern_t ^ mask.reshape(self.pattern_t.shape)
@@ -374,7 +374,7 @@ class Probe:
             tag = f"rand_q{q}"
             if self.skip("tieflip", tag):
                 continue
-            n_flip = int(round(q * self.pattern_t.size))
+            n_flip = round(q * self.pattern_t.size)
             mask = np.zeros(self.pattern_t.size, dtype=bool)
             mask[rng.choice(self.pattern_t.size, n_flip, replace=False)] = True
             pat = self.pattern_t ^ mask.reshape(self.pattern_t.shape)
